fix: u8 bug

+++ Library/PAX_SAPIENTICA/Simulation/KanakumaLifeSpan.hpp
@@ -47,48 +47,48 @@ namespace paxs {
 
         /// @brief 英語未翻訳
         /// @brief 寿命を決定する
-        std::uint_least8_t setLifeSpan(const std::uint_least8_t gender_, std::mt19937& gen) {
+        std::uint_least32_t setLifeSpan(const std::uint_least8_t gender_, std::mt19937& gen) {
 
             if (life_person_num(gen) <= 37) { // もし子供だったら
-                const std::uint_least8_t child_type = static_cast<std::uint_least8_t>(life_child_num(gen));
+                const std::uint_least32_t child_type = static_cast<std::uint_least32_t>(life_child_num(gen));
 
                 if (child_type <= 6) { // 乳児
-                    return static_cast<std::uint_least8_t>(life_infant_exp_dist(gen));
+                    return static_cast<std::uint_least32_t>(life_infant_exp_dist(gen));
                 }
                 else if (child_type <= (6 + 22)) { // 幼児
-                    return static_cast<std::uint_least8_t>(life_toddler_exp_dist(gen));
+                    return static_cast<std::uint_least32_t>(life_toddler_exp_dist(gen));
                 }
                 else if (child_type <= (6 + 22 + 5)) { // 小児
-                    return static_cast<std::uint_least8_t>(life_child_exp_dist(gen));
+                    return static_cast<std::uint_least32_t>(life_child_exp_dist(gen));
                 }
                 // 若年
-                return static_cast<std::uint_least8_t>(life_young_exp_dist(gen));
+                return static_cast<std::uint_least32_t>(life_young_exp_dist(gen));
             }
             // もし大人だったら
             if (gender_ == 0) { // 女性の場合
-                const std::uint_least8_t adult_type = static_cast<std::uint_least8_t>(life_male_adult_num(gen));
+                const std::uint_least32_t adult_type = static_cast<std::uint_least32_t>(life_male_adult_num(gen));
 
                 if (adult_type <= 14) { // 成年
-                    return static_cast<std::uint_least8_t>(life_adult_exp_dist(gen));
+                    return static_cast<std::uint_least32_t>(life_adult_exp_dist(gen));
                 }
                 else if (adult_type <= (14 + 26)) { // 熟年
-                    return static_cast<std::uint_least8_t>(life_mature_exp_dist(gen));
+                    return static_cast<std::uint_least32_t>(life_mature_exp_dist(gen));
                 }
                 // 老年
-                return static_cast<std::uint_least8_t>(life_older_exp_dist(gen));
+                return static_cast<std::uint_least32_t>(life_older_exp_dist(gen));
             }
             else if (gender_ == 1) { // 男性の場合
-                const std::uint_least8_t adult_type = static_cast<std::uint_least8_t>(life_female_adult_num(gen));
+                const std::uint_least32_t adult_type = static_cast<std::uint_least32_t>(life_female_adult_num(gen));
 
                 if (adult_type <= 19) { // 成年
-                    return static_cast<std::uint_least8_t>(life_adult_exp_dist(gen));
+                    return static_cast<std::uint_least32_t>(life_adult_exp_dist(gen));
                 }
                 else if (adult_type <= (19 + 19)) { // 熟年
-                    return static_cast<std::uint_least8_t>(life_mature_exp_dist(gen));
+                    return static_cast<std::uint_least32_t>(life_mature_exp_dist(gen));
                 }
             }
             // 老年
-            return static_cast<std::uint_least8_t>(life_older_exp_dist(gen));
+            return static_cast<std::uint_least32_t>(life_older_exp_dist(gen));
         }
 
     };
